+++ connectors/random/random.go
@@ -184,7 +184,7 @@ func (c *Connector) CheckSchemaStatus(ctx context.Context, scope, namePrefix str
 }
 
 // CreateScope returns success
-func (c *Connector) CreateScope(ctx context.Context, scope string) error {
+func (c *Connector) CreateScope(ctx context.Context, _ *dosa.ScopeMetadata) error {
 	return nil
 }